SPIKE: Having a go at tag limits.
This is a spike and does not pass the tests. This is for review/discussion only.

I have had a go at getting the tag limits working but i'm getting to the point
where I am a bit stuck and need some help.

I tried to introduce the concept of the `test_suite` as a first class object
that the compiler knows about so that visitors know when the test suit has
started/finished. Knowing when the test_suite has started/finished allows for
a tag counter to collect the tags and allow the enforcement of the limits when
the compiler has run through the test suite.

Am I heading in the right direction or am I miles off?

What are your opinions @mattwynne, @tooky, @os97673?